requirements.txt
absl-py

aiohttp
aiosignal
albumentations
annotated-types
antlr4-python3-runtime
anyio
appdirs
asttokens
attrs
accelerate
bitsandbytes
blinker
blis
bokeh
braceexpand
Brotli
cachetools
catalogue
certifi
cffi
charset-normalizer
click
cloudpathlib
colorama
comm
confection
contourpy
cryptography
cycler
cymem
dataclasses
datasets
debugpy
decorator
decord
dill
distro
docker-pycreds
einops
einops-exts

et-xmlfile
exceptiongroup
executing
filelock
fonttools
frozenlist
fsspec
ftfy
gitdb
GitPython
google-auth
google-auth-oauthlib
grad-cam
grpcio
h11
h2
hpack
httpcore
httpx
huggingface_hub
hydra-core
hyperframe
idna
imagecodecs
imageio
importlib_metadata
importlib_resources
inflection
iopath
ipykernel
ipython
jedi
Jinja2
joblib
jupyter_client
jupyter_core
kiwisolver
langcodes
lazy_loader
lightning-bolts
lightning-utilities
llvmlite
Markdown
MarkupSafe
matplotlib
matplotlib-inline
more-itertools
mpmath
multidict
multiprocess
munkres
murmurhash
nest_asyncio
networkx
nibabel
nltk
numba
numpy
oauthlib
omegaconf
open_clip_torch
open-flamingo
openai
opencv-python
openpyxl
packaging
pandas
parso
pathtools
patsy
peft
pexpect
pickleshare
pillow
pip
platformdirs
ply
portalocker
preshed
prompt-toolkit
protobuf
psutil
ptyprocess
pure-eval
pyarrow
pyasn1
pyasn1-modules
pycparser
pydantic
pydantic_core
pyDeprecate
Pygments
PyJWT
pynndescent
pyOpenSSL
pyparsing
PyQt5
PyQt5-sip
PySocks
python-dateutil
pytorch-lightning
pytz
pyu2f
PyWavelets
PyYAML
pyzmq
qudida
regex
requests
requests-oauthlib
rsa
safetensors
scikit-image
scikit-learn
scipy
seaborn
sentencepiece
sentry-sdk
setproctitle
setuptools
sip
six
smart-open
smmap
sniffio
spacy
spacy-legacy
spacy-loggers
srsly
stack-data
statsmodels
sympy
tensorboard
tensorboard-data-server
thinc
threadpoolctl
tifffile
timm
tokenizers
toml
tomli
torch
torchmetrics
torchvision
tornado
tqdm
traitlets
transformers
ttach
typer
typing_extensions
tzdata
umap-learn
urllib3
wandb
wasabi
wcwidth
weasel
webdataset
Werkzeug
wheel
xxhash
xyzservices
yarl
zipp